+++ packages/popmotion/src/action/index.ts
@@ -1,6 +1,6 @@
 import Chainable from '../chainable';
 import createObserver from '../observer';
-import { ObserverCandidate } from '../observer/types';
+import { ObserverCandidate, PartialObserver } from '../observer/types';
 import { ActionInit, ActionProps, ColdSubscription } from './types';
 
 export class Action extends Chainable<Action> {
@@ -8,7 +8,7 @@ export class Action extends Chainable<Action> {
     return new Action(props);
   }
 
-  start(observerCandidate: ObserverCandidate): ColdSubscription {
+  start(observerCandidate: ObserverCandidate = {}): ColdSubscription {
     const { init, ...observerProps } = this.props;
     const observer = createObserver(observerCandidate, observerProps);
     const api = init(observer);
@@ -20,6 +20,10 @@ export class Action extends Chainable<Action> {
     const subscription = api
       ? { ...defaultSubscription, ...api }
       : defaultSubscription;
+
+    if ((observerCandidate as PartialObserver).registerParent) {
+      (observerCandidate as PartialObserver).registerParent(subscription);
+    }
 
     return subscription;
   }