+++ packages/popmotion/src/animations/physics/_tests/physics.test.ts
@@ -9,7 +9,7 @@ describe('physics', () => {
       }).start((v) => i = v);
       setTimeout(() => {
         a.stop();
-        (i >= 100 && i <= 110) ? resolve() : reject();
+        (i >= 100 && i <= 110) ? resolve() : reject(i);
       }, 1000);
     });
   });